Redesign service so that serialization is handled by ISplitsFacade instead of by IGameManager.
Serialization and deserialization should be handled outside of GameManager since clients should be allowed to load files without know what game the file is for. Clients can also choose whether or not a loaded game should be tracked in the GameManager's RecentSplits list (which should probably be renamed to FavoritesList or something).